+++ routers/common.go
@@ -1,10 +1,13 @@
-package commonRouters
+package routers
 
 import (
-	"github.com/gin-gonic/gin"
 	"net/http"
+	"runtime"
 	"swan-provider/common"
 	"swan-provider/common/constants"
+	"swan-provider/models"
+
+	"github.com/gin-gonic/gin"
 )
 
 func HostManager(router *gin.RouterGroup) {
@@ -14,4 +17,13 @@ func HostManager(router *gin.RouterGroup) {
 func GetSwanMinerVersion(c *gin.Context) {
 	info := getSwanMinerHostInfo()
 	c.JSON(http.StatusOK, common.CreateSuccessResponse(info))
+}
+
+func getSwanMinerHostInfo() *models.HostInfo {
+	info := new(models.HostInfo)
+	info.SwanMinerVersion = common.GetVersion()
+	info.OperatingSystem = runtime.GOOS
+	info.Architecture = runtime.GOARCH
+	info.CPUnNumber = runtime.NumCPU()
+	return info
 }
