--- routers/commonRouters/commonService.go
@@ -1,16 +1,0 @@
-package commonRouters
-
-import (
-	"runtime"
-	"swan-provider/common"
-	"swan-provider/models"
-)
-
-func getSwanMinerHostInfo() *models.HostInfo {
-	info := new(models.HostInfo)
-	info.SwanMinerVersion = common.GetVersion()
-	info.OperatingSystem = runtime.GOOS
-	info.Architecture = runtime.GOARCH
-	info.CPUnNumber = runtime.NumCPU()
-	return info
-}
